+++ NEWS.md
@@ -3,6 +3,7 @@
 * Added codecov to measure test coverage
 * Added AppVeyor continuous integration
 * Added option not only to join based on a maximum distance but also to append a distance column, using the `distance_col` argument. This is available in `difference_join`, `distance_join`, `stringdist_join`, and `geo_join`
+* Fixed to ignore groups while preserving the groups of x in the output
 
 # fuzzyjoin 0.1